Pull native library loading helpers out of dllimport.cpp (direct copy)
